Round-2 review fix wave A+B: export integrity + training math (5-lens re-review, ~24 findings)
Export integrity (A):
- reconvert_onnx: construct CeresNet from config loss multipliers (hardcoded
  zeros silently exported value2 as a COPY of value1 for dual-value nets);
  universal key-set guard + strict load, mirroring recover_export
- delete recover_export_hostname.py (redundant with recover_export.py +
  CERES_HOST_PREFIX; had strict=False and none of the corruption guards)
- config_bootstrap: bridge FFNSoftCap/RopeBase (parameter-free forward-math
  env knobs invisible to the key-set guard - fresh-shell re-export built a
  different function than trained)
- train: correct the mathematically wrong EMA-under-DDP guidance (one step
  consumes the same global positions regardless of world size - use the
  single-GPU-validated EMAPeriodSteps UNCHANGED); assert forward-batch
  divisibility by world_size

Training math + fp16 serving (B):
- fp16 export: max_finite_val 1e4 -> 65504 (no more silent weight clamping a
  decade below the fp16 ceiling); block ReduceLogSumExp from fp16 so the
  dual-plane soft-min islands serve in fp32 as trained
- BFloat16Pure: cast an export copy to fp32 before ONNX trace (export could
  never succeed; error was swallowed by the blanket except)
- QK-clip stash max-accumulates across the accumulation window (was
  last-micro-batch only - 7/8 of spikes invisible to tau); apply resets
- fp32 guards on policy/action/value-diff CE (BFloat16Pure parity with the
  explicitly guarded value paths)
- smolref: softcap the reference map too; HLGaussSigmaScale explicit-0
  respected (falsy-swallow); 4-board action-loss None guards; SOAP debias
  uses the asserted uniform step (loop-leak gave ZeroDivision on late-init
  params); AdEMAMix bias-corrects only m1 per the paper; SoftMoE x-norm on
  the feature axis (was token axis)

Validated: reconvert end-to-end on f3dpgiC (ORT load, value1!=value2,
bit-identical outputs vs the train.py export); 2-GPU DDP smoke 2M pos clean
(0 NaN, clip-sync collective exercised, corrected EMA banner on rank 0)

+++ src/CeresTrainPy/losses.py
@@ -346,7 +346,10 @@ class LossCalculator():
       # for cross-run comparability; only the RETURNED loss is reweighted.
       _log_loss = _ce_i.mean() - entropy
     else:
-      loss = self.ce_loss.forward(output, target) - entropy
+      # fp32-vern (runde-2-funn): under BFloat16Pure (ingen autocast) ville
+      # log_softmax over 1858 logits ellers regnes i bf16 — value-stiene har
+      # dette vernet eksplisitt (_prov_weighted_ce), policy manglet det.
+      loss = self.ce_loss.forward(output.float(), target.float()) - entropy
       _log_loss = loss
 
     self.PENDING_POLICY_LOSS += _log_loss.item() if not calc_grad_norm_mode else 0
@@ -542,9 +545,9 @@ class LossCalculator():
     if calc_grad_norm_mode:
       self.model.zero_grad()
 
-    target_softmax = F.softmax(target.detach(), dim=-1)  # maalsiden skal ikke motta gradient (bugfunn 2026-08-28)
+    target_softmax = F.softmax(target.detach().float(), dim=-1)  # maalsiden skal ikke motta gradient (bugfunn 2026-08-28); fp32 = BFloat16Pure-vern
     entropy = self.entropy(target_softmax) if subtract_entropy else 0.0
-    loss = self.ce_loss.forward(output, target_softmax) - entropy
+    loss = self.ce_loss.forward(output.float(), target_softmax) - entropy
 
     self.PENDING_VALUE_DIFF_LOSS += loss.item() if not calc_grad_norm_mode else 0
     return self.calc_loss_grad_norm('value_diff', loss, loss_wt) if calc_grad_norm_mode else loss
@@ -554,9 +557,9 @@ class LossCalculator():
     if calc_grad_norm_mode:
       self.model.zero_grad()
 
-    target_softmax = F.softmax(target.detach(), dim=-1)  # maalsiden skal ikke motta gradient (bugfunn 2026-08-28)
+    target_softmax = F.softmax(target.detach().float(), dim=-1)  # maalsiden skal ikke motta gradient (bugfunn 2026-08-28); fp32 = BFloat16Pure-vern
     entropy = self.entropy(target_softmax) if subtract_entropy else 0.0
-    loss = self.ce_loss(output, target_softmax) - entropy
+    loss = self.ce_loss(output.float(), target_softmax) - entropy
    
     self.PENDING_VALUE2_DIFF_LOSS += loss.item() if not calc_grad_norm_mode else 0
     return self.calc_loss_grad_norm('value2_diff', loss, loss_wt) if calc_grad_norm_mode else loss
@@ -566,9 +569,9 @@ class LossCalculator():
     if calc_grad_norm_mode:
       self.model.zero_grad()
 
-    target_softmax = F.softmax(target, dim=-1)
+    target_softmax = F.softmax(target.float(), dim=-1)
     entropy = self.entropy(target_softmax) if subtract_entropy else 0.0
-    loss = self.ce_loss(output, target_softmax) - entropy
+    loss = self.ce_loss(output.float(), target_softmax) - entropy
   
     self.PENDING_ACTION_LOSS += loss.item() if not calc_grad_norm_mode else 0
     return self.calc_loss_grad_norm('action', loss, loss_wt) if calc_grad_norm_mode else loss
